Add missing docstring and move classification out of benchmark module
