+++ ci/tests/test_benchmark/test_benchmark.py
@@ -1,6 +1,6 @@
 from helical.benchmark.benchmark import Benchmark
-from helical.benchmark.classification.classifier import Classifier
-from helical.benchmark.classification.neural_network import NeuralNetwork
+from helical.classification.classifier import Classifier
+from helical.classification.neural_network import NeuralNetwork
 from helical.models.scgpt.model import scGPT
 import anndata as ad
 import numpy as np
