+++ examples/benchmark.py
@@ -1,12 +1,12 @@
 from helical.benchmark.benchmark import Benchmark
 from helical.models.geneformer.model import Geneformer
 from helical.models.scgpt.model import scGPT
-from helical.benchmark.classification.neural_network import NeuralNetwork
-from helical.benchmark.classification.svm import SupportVectorMachine as SVM
+from helical.classification.neural_network import NeuralNetwork
+from helical.classification.svm import SupportVectorMachine as SVM
+from helical.classification.classifier import Classifier
 import anndata as ad
 from omegaconf import DictConfig
 import hydra
-from helical.benchmark.classification.classifier import Classifier
 
 @hydra.main(version_base=None, config_path=".", config_name="config")
 def benchmark(cfg: DictConfig) -> None:
@@ -23,7 +23,7 @@ def benchmark(cfg: DictConfig) -> None:
     # saved_svm_head = SVM().load('my_svm.pkl')
     # scgpt_loaded_svm = Classifier().load_model(scgpt, saved_svm_head, "scgpt with saved SVM")           
 
-    gene_c = Classifier().train_classifier_head(train_data, geneformer, NeuralNetwork(**cfg["neural_network"]), 0.2, 42)
+    gene_c = Classifier().train_classifier_head(train_data, geneformer, NeuralNetwork(**cfg["neural_network"]), "cell_type", 0.2, 42)
     scgpt_nn_c = Classifier().train_classifier_head(train_data, scgpt, NeuralNetwork(**cfg["neural_network"]))           
 
     bench = Benchmark()
